Revert "moved background to style sheet"
This reverts commit 57843551773507d9bba072b0548c0c0415ad6a3b.

I was play ing around and broke the formatting so I reverted back to a previous state.

As for changed i helped wrap the content around the paragraph a little better.

I also adjusted your media queries so the image in your paragraph is nested on top instead of off to the side. It was a tad out of frame.

I also adjusted the font styling, color and alignment so it could be read easier

These changes are based off my own opinion an in no way needed, you did a great job!